requirements.txt
imutils==0.5.2
numpy==1.19.2
opencv-python==4.4.0.44
torch==1.8.1
torchvision==0.9.1

clams-python==0.5.1